+++ validator.js
@@ -1353,7 +1353,7 @@ function normalizeEmail(email, options) {
   return parts.join('@');
 }
 
-var version = '8.1.0';
+var version = '8.2.0';
 
 var validator = {
   version: version,